+++ KiLib/Raster/Raster_DEM.cpp
@@ -1,0 +1,130 @@
+#include <KiLib/Raster/Raster.hpp>
+#include <spdlog/fmt/ostr.h>
+#include <spdlog/spdlog.h>
+
+namespace KiLib
+{
+   void Raster::fromDEM(const std::string path)
+   {
+      std::ifstream      rasterFile;
+      std::string        line, key;
+      std::istringstream stream(line);
+
+      // Open Elevation file
+      rasterFile.open(path, std::ios::in);
+      if (!rasterFile.is_open())
+      {
+         spdlog::error("Failed to open {} for reading", path);
+         exit(EXIT_FAILURE);
+      }
+
+      // Load header
+      for (int i = 0; i < 6; i++)
+      {
+         getline(rasterFile, line);
+         stream.str(line);
+         stream.clear();
+
+         stream >> key;
+
+         // Force lower-case keys
+         std::transform(key.begin(), key.end(), key.begin(), ::tolower);
+
+         if (key == "ncols")
+         {
+            stream >> this->nCols;
+         }
+         else if (key == "nrows")
+         {
+            stream >> this->nRows;
+         }
+         else if (key == "xllcorner")
+         {
+            stream >> this->xllcorner;
+         }
+         else if (key == "yllcorner")
+         {
+            stream >> this->yllcorner;
+         }
+         else if (key == "cellsize")
+         {
+            stream >> this->cellsize;
+         }
+         else if (key == "nodata_value")
+         {
+            stream >> this->nodata_value;
+         }
+         else
+         {
+            std::cerr << "Unexpected value in header!" << std::endl;
+            return;
+         }
+      }
+
+      this->data.resize(nRows * nCols);
+      // Load elevations
+      for (size_t row = 0; row < this->nRows; row++)
+      {
+         getline(rasterFile, line);
+         stream.str(line);
+         stream.clear();
+
+         // Moving along a row (across columns) is movement through X
+         // Moving down a column (across rows) is movement through Y
+         for (size_t col = 0; col < this->nCols; col++)
+         {
+            double value;
+            stream >> value;
+            this->at(this->nRows - row - 1, col) = value;
+         }
+      }
+
+      this->width       = (this->nCols - 1) * this->cellsize;
+      this->height      = (this->nRows - 1) * this->cellsize;
+      this->nData       = this->nRows * this->nCols;
+      this->constructed = true;
+
+      rasterFile.close();
+   }
+
+
+   void Raster::toDEM(const std::string path) const
+   {
+      std::ofstream outFile = std::ofstream(path);
+      if (!outFile.is_open())
+      {
+         spdlog::error("Cannot open output {}", path);
+         exit(EXIT_FAILURE);
+      }
+
+      fmt::print(
+         outFile,
+         "ncols {}\n"
+         "nrows {}\n"
+         "xllcorner {:.4f}\n"
+         "yllcorner {:.4f}\n"
+         "cellsize {:.4f}\n"
+         "NODATA_value {}\n",
+         this->nCols, this->nRows, this->xllcorner, this->yllcorner, this->cellsize, this->nodata_value);
+
+      for (size_t row = 1; row <= this->nRows; row++)
+      {
+         for (size_t col = 0; col < this->nCols; col++)
+         {
+            double val = this->operator()(this->nRows - row, col);
+
+            if (val == this->nodata_value)
+            {
+               fmt::print(outFile, "{} ", this->nodata_value);
+            }
+            else
+            {
+               fmt::print(outFile, "{: .8f} ", val);
+            }
+         }
+         fmt::print(outFile, "\n");
+      }
+
+      outFile.close();
+   }
+} // namespace KiLib
